Language/Python/no-gil/benchmarks/requirements.txt
throttled-py==1.0.2
pyperformance==1.11.0
numpy==2.2.4
